[#23] code clean-up and refactor
