+++ YoutubeQueuerJS/server/config.js
@@ -1,0 +1,5 @@
+var config = {
+    port: 8080
+};
+
+module.exports = config;
