+++ YoutubeQueuerJS/server/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+
+<html>
+
+<body>
+    <div>
+        <button id="signInButton">Authorize app for Youtube access </button>
+    </div>
+    <script src="/scripts/index.js"></script>
+</body>
+
+</html>
